+++ src/CMakeLists.txt
@@ -37,6 +37,6 @@ if (MUPEN64RR_BUILD_WIN32)
     add_subdirectory(Plugins.Win32.TASAudio)
     add_subdirectory(Plugins.Win32.TASRSP)
 
-    # New xplat plugins
-    add_subdirectory(Plugins.Win32.SDLAudio)
+    # Legacy plugins
+    add_subdirectory(Plugins.Win32.TASAudioLegacy)
 endif()
