--- src/Plugins.Win32.SDLAudio/Resource.h
@@ -1,7 +1,0 @@
-#pragma once
-
-#define IDD_CONFIG 101
-
-#define IDC_SWAP_CHANNELS 1002
-#define IDC_SYNC_AUDIO 1003
-
