--- src/Plugins.Win32.SDLAudio/Resource.rc
@@ -1,36 +1,0 @@
-#include "resource.h"
-#include <Windows.h>
-
-#define IDC_STATIC -1
-
-// VERTICAL LAYOUT DEFINITIONS
-// =====================================================
-
-#define DDX_PLAYBACK_Y 5
-#define DDX_PLAYBACK_IY (DDX_PLAYBACK_Y + 15)
-#define DDX_PLAYBACK_H 52
-
-#define DDX_DBTNS_Y (DDX_PLAYBACK_Y + DDX_PLAYBACK_H + 6)
-#define DDX_DBTNS_H 14
-
-#define DDX_DIALOG_H (DDX_DBTNS_Y + DDX_DBTNS_H + 6)
-
-// DIALOG_DEFINITION
-// =====================================================
-
-IDD_CONFIG DIALOGEX 0, 0, 200, DDX_DIALOG_H
-STYLE DS_SETFONT | DS_MODALFRAME | DS_CENTER | WS_POPUP | WS_CAPTION | WS_SYSMENU
-CAPTION "SDL Audio Settings"
-LANGUAGE LANG_ENGLISH, SUBLANG_ENGLISH_US
-FONT 8, "MS Shell Dlg"
-BEGIN
-    // EDIT "Default"
-    CONTROL "Playback options", IDC_STATIC, BUTTON, BS_GROUPBOX, 10, DDX_PLAYBACK_Y, 180, DDX_PLAYBACK_H
-
-    CONTROL "Swap channels", IDC_SWAP_CHANNELS, BUTTON, BS_AUTOCHECKBOX | WS_TABSTOP, 16, (DDX_PLAYBACK_IY + 0), 116, 9
-    CONTROL "Sync audio", IDC_SYNC_AUDIO, BUTTON, BS_AUTOCHECKBOX | WS_TABSTOP, 16, (DDX_PLAYBACK_IY + 14), 116, 9
-
-
-    PUSHBUTTON "Cancel", IDCANCEL, 87, DDX_DBTNS_Y, 50, DDX_DBTNS_H
-    DEFPUSHBUTTON "&OK", IDOK, 140, DDX_DBTNS_Y, 50, DDX_DBTNS_H
-END
