+++ src/Plugins.Win32.TASAudio/Config.cpp
