+++ src/Plugins.Win32.TASAudio/Main_Win32.cpp
