+++ src/Plugins.Win32.TASAudio/Resource.h
@@ -1,28 +1,7 @@
-//{{NO_DEPENDENCIES}}
-// Microsoft Visual C++ generated include file.
-// Used by resource.rc
-//
-#define IDD_CONFIG 101
-#define IDC_AI 1001
-#define IDC_OLDSYNC 1002
-#define IDC_VOLUME 1003
-#define IDC_DISALLOWDS8 1004
-#define IDC_DISALLOWXA2 1007
-#define IDC_DEVICE 1017
-#define IDC_SLIDER_BUFFERFPS 1021
-#define IDC_SLIDER_BACKFPS 1022
-#define IDC_BUFFERS 1024
-#define IDC_BUFFERS_TEXT 1025
-#define IDC_SLIDER_BACKFPS_TEXT 1026
-#define IDC_SLIDER_BUFFERFPS_TEXT 1027
+#pragma once
 
-// Next default values for new objects
-//
-#ifdef APSTUDIO_INVOKED
-#ifndef APSTUDIO_READONLY_SYMBOLS
-#define _APS_NEXT_RESOURCE_VALUE 104
-#define _APS_NEXT_COMMAND_VALUE 40001
-#define _APS_NEXT_CONTROL_VALUE 1026
-#define _APS_NEXT_SYMED_VALUE 101
-#endif
-#endif
+#define IDD_CONFIG 101
+
+#define IDC_SWAP_CHANNELS 1002
+#define IDC_SYNC_AUDIO 1003
+
