+++ src/Plugins.Win32.TASAudioLegacy/ABI1.cpp
