+++ src/Plugins.Win32.TASAudioLegacy/ABI2.cpp
