+++ src/Plugins.Win32.TASAudioLegacy/ABI3.cpp
