+++ src/Plugins.Win32.TASAudioLegacy/AudioHLE.h
