+++ src/Plugins.Win32.TASAudioLegacy/Configuration.cpp
