+++ src/Plugins.Win32.TASAudioLegacy/DirectSoundDriver.h
