+++ src/Plugins.Win32.TASAudioLegacy/HLEMain.cpp
