+++ src/Plugins.Win32.TASAudioLegacy/Mupen64plusHLE/Mupen64Support.cpp
