+++ src/Plugins.Win32.TASAudioLegacy/Mupen64plusHLE/audio.h
