+++ src/Plugins.Win32.TASAudioLegacy/Mupen64plusHLE/hle_internal.h
