+++ src/Plugins.Win32.TASAudioLegacy/SoundDriver.cpp
