+++ src/Plugins.Win32.TASAudioLegacy/SoundDriver.h
