+++ src/Plugins.Win32.TASAudioLegacy/SoundDriverInterface.cpp
