+++ src/Plugins.Win32.TASAudioLegacy/Types.h
